SJ_FIX_LICENSE_AND_FIX_LONG_TYPE

+++ FlexDelivery/src/main/java/com/project/fd/owner/ownerregister/model/OwnerRegisterVO.java
@@ -1,7 +1,7 @@
 package com.project.fd.owner.ownerregister.model;
 
 public class OwnerRegisterVO {
-	private int oRegisterNo; /* 사업자등록번호 */
+	private long oRegisterNo; /* 사업자등록번호 */
 	private String oRegisterFileName; /* 사업자등록증파일이름 */
 	private String oRegisterOriginalFileName;/* 사업자등록증원본파일이름 */
 	private int aAgreeNo;  /* 관리자승인번호 fk */
@@ -18,10 +18,10 @@ public class OwnerRegisterVO {
 	public void setOwnerregisterRegdate(String ownerregisterRegdate) {
 		this.ownerregisterRegdate = ownerregisterRegdate;
 	}
-	public int getoRegisterNo() {
+	public long getoRegisterNo() {
 		return oRegisterNo;
 	}
-	public void setoRegisterNo(int oRegisterNo) {
+	public void setoRegisterNo(long oRegisterNo) {
 		this.oRegisterNo = oRegisterNo;
 	}
 	public String getoRegisterFileName() {
